sqlserver/requirements.txt
Faker
pyodbc
python-decouple
python-dotenv
